api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kulunchick", email = "nikolikrasovski@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn>=0.32.0",
    "grpcio>=1.67.1",
    "grpc-interceptor>=0.15.4",
    "grpcio-tools>=1.67.1",
    "pydantic-settings>=2.6.1",
    "logfire[fastapi]>=2.3.0",
    "dishka>=1.4.1",
    "pytest-asyncio>=0.25.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
virtual = true
managed = true
dev-dependencies = [
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "types-protobuf>=5.29.1.20241207",
    "pytest>=8.3.4",
    "uuid7>=0.1.0",
    "httpx>=0.28.1",
]

[tool.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "C901",  # too complex
]
